+++ app/src/main/java/com/auf/cea/openweatherapilesson/models/currentapi/Sys.kt
@@ -3,7 +3,7 @@ package com.auf.cea.openweatherapilesson.models.currentapi
 data class Sys(
     var country: String,
     var id: Int,
-    var sunrise: Int,
-    var sunset: Int,
+    var sunrise: Long,
+    var sunset: Long,
     var type: Int
 )
